==> scripts/01_reading_large.ipynb <==
# import zipfile

# def unzip_file(zip_path, extract_to):
#     with zipfile.ZipFile(zip_path, 'r') as zip_ref:
#         zip_ref.extractall(extract_to)

# zip_path = '..\\raw_data\\randhrs1992_2020v2_STATA.zip'
# extract_to = '..\\raw_data'

# unzip_file(zip_path, extract_to)

import psutil

# Get system memory information
memory_info = psutil.virtual_memory()

# Display memory information
print(f"Total memory: {memory_info.total / (1024 ** 3):.2f} GB")
print(f"Available memory: {memory_info.available / (1024 ** 3):.2f} GB")
print(f"Used memory: {memory_info.used / (1024 ** 3):.2f} GB")
print(f"Memory usage: {memory_info.percent}%")

import pyreadstat

df, meta = pyreadstat.read_dta('..\\raw_data\\randhrs1992_2020v2.dta') # 7min with adequate memory

num_subsets = 50

rows_per_subset = len(df) // num_subsets

for i in range(num_subsets):
    start_row = i * rows_per_subset
    end_row = start_row + rows_per_subset
    
    # Handle the last subset to include any remaining rows
    if i == num_subsets - 1:
        subset_df = df.iloc[start_row:]
    else:
        subset_df = df.iloc[start_row:end_row]
    
    subset_df.to_csv(f'..\\intermediate_data\\wide\\subset_{i+1}.csv', index=False)
