==> scripts/02_filtering_subsets.ipynb <==
import pandas as pd
from variables_dict_builder import variables_dict_builder

variables_dict = variables_dict_builder()
variables_suffixes = [*variables_dict.keys()]

num_subsets = 50

for i in range(num_subsets):
    temp_df = pd.read_csv(f'../intermediate_data/all_columns/subset_{i+1}.csv')
    # Selecting only the columns for which we define above and saving
    temp_df.filter(regex='|'.join([f'{suffix}$' for suffix in variables_suffixes])).to_csv(f'../intermediate_data/filtered_columns/filtered_{i+1}.csv')
