==> scripts/03_stacking_filtereds.ipynb <==
import pandas as pd
import os

FILTERED_DATA_DIR = '../intermediate_data/filtered_columns'

data_frames = []

for i in range(1, 51):  # Loop from 1 to 50
    file_path = os.path.join(FILTERED_DATA_DIR, f'filtered_{i}.csv')
    temp_df = pd.read_csv(file_path)
    data_frames.append(temp_df)

stacked_df = pd.concat(data_frames, ignore_index=True)

stacked_df.to_csv('../intermediate_data/stacked_filtered_data.csv', index=False)
