==> scripts/07_summary_statistics.ipynb <==
import pandas as pd

df = pd.read_csv('../final_data/clean_model_data.csv')

black_flag = 1
hispan_flag = 0

(df[(df['rablack']==black_flag) & (df['rahispan']==hispan_flag)]
    ['rafeduc']
    .agg(['mean','std','min','max'])
)
