==> scripts/99_browsing_subset.ipynb <==
import pandas as pd

peek_df = pd.read_csv('../intermediate_data/all_columns/subset_1.csv')


variable_suffix = "age"
[col for col in peek_df.columns if variable_suffix in col]

# peek_df[['hhidpn','r2beda']]
